feat: add hydrate fallback component for loading states
feat: implement NotFound component for 404 error pages

feat: create UserCard component for displaying user information

feat: establish AuthContext for user authentication management

feat: create useAuth hook for accessing authentication context

feat: implement useSocket hook for managing WebSocket connections

feat: add utility functions for user and channel management

feat: create ChannelInfoModal for displaying channel details and managing members

feat: implement CreateInvite component for generating invite links

feat: create EditMessage component for editing messages in chat

feat: add AddFriend component for sending friend requests

feat: implement Friends page to manage friend lists and requests

feat: create Invites page for handling channel invites

feat: implement NotFound page routing for undefined routes

feat: add avatar generation script for users without avatars

test: add unit tests for FriendsController

feat: implement FriendsController for managing friend requests

feat: create DTOs for friend request handling

feat: set up FriendsModule for friendship management

test: add unit tests for FriendsService

feat: implement InvitesController for managing invites

feat: create InvitesService for invite link handling

feat: set up InvitesModule for invite management

+++ apps/server/src/modules/channels/channels.service.ts
@@ -1,13 +1,25 @@
 import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import * as crypto from 'crypto';
 
 import { QueryService } from '~/common/services/query.service';
 import { Channel, Prisma, PrismaService } from '~/database';
 
+import { FriendsService } from '../friendships/friends.service';
 import { CreateChannelDto, QueryChannelDto, UpdateChannelDto } from './channels.dto';
 
 @Injectable()
 export class ChannelsService extends QueryService<Channel> {
-  constructor(private prisma: PrismaService) {
+  userSelect: Prisma.UserSelect = {
+    avatarUrl: true,
+    displayName: true,
+    id: true,
+    username: true,
+  };
+
+  constructor(
+    private prisma: PrismaService,
+    private friendsService: FriendsService,
+  ) {
     super(prisma.channel);
   }
 
@@ -27,14 +39,76 @@ export class ChannelsService extends QueryService<Channel> {
     return channel;
   }
 
-  async findAll(query: QueryChannelDto, userId: string) {
-    const where: Prisma.ChannelWhereInput = {
-      users: {
-        some: { id: userId },
-      },
-    };
+  async createDMChannel(userId: string, targetUserId: string) {
+    const isFriend = await this.friendsService.isFriend(userId, targetUserId);
 
-    const channels = await this.queryAll(query, ['name', 'description'], where);
+    if (!isFriend) {
+      throw new ForbiddenException('You can only create a DM channel with friends.');
+    }
+
+    const existingChannel = await this.prisma.channel.findFirst({
+      where: {
+        type: 'DM',
+        users: {
+          every: {
+            id: { in: [userId, targetUserId] },
+          },
+        },
+      },
+    });
+
+    if (existingChannel) {
+      return existingChannel;
+    }
+
+    const dmChannel = await this.prisma.channel.create({
+      data: {
+        isPublic: false,
+        name: crypto.randomBytes(16).toString('hex'),
+        owner: {
+          connect: { id: userId },
+        },
+        type: 'DM',
+        users: {
+          connect: [{ id: userId }, { id: targetUserId }],
+        },
+      },
+    });
+
+    return dmChannel;
+  }
+
+  async findAll(query: QueryChannelDto, userId: string) {
+    console.log(query);
+
+    const channels = await this.prisma.channel.findMany({
+      include: {
+        users: { select: this.userSelect },
+      },
+      where: {
+        users: {
+          some: { id: userId },
+        },
+      },
+    });
+
+    return channels;
+  }
+
+  async findAllDMChannels(userId: string) {
+    const channels = await this.prisma.channel.findMany({
+      include: {
+        users: {
+          select: this.userSelect,
+        },
+      },
+      where: {
+        type: 'DM',
+        users: {
+          some: { id: userId },
+        },
+      },
+    });
 
     return channels;
   }
@@ -42,7 +116,13 @@ export class ChannelsService extends QueryService<Channel> {
   async findMessages(id: string, userId: string) {
     const channel = await this.prisma.channel.findFirst({
       include: {
-        messages: true,
+        messages: {
+          include: {
+            author: {
+              select: this.userSelect,
+            },
+          },
+        },
       },
       where: {
         id,
@@ -59,6 +139,9 @@ export class ChannelsService extends QueryService<Channel> {
 
   async findOne(id: string, userId: string) {
     const channel = await this.prisma.channel.findFirst({
+      include: {
+        users: { select: this.userSelect },
+      },
       where: {
         id,
         users: { some: { id: userId } },
@@ -72,11 +155,63 @@ export class ChannelsService extends QueryService<Channel> {
     return channel;
   }
 
+  async kickUserFromChannel(channelId: string, userId: string, targetUserId: string) {
+    const channel = await this.findOne(channelId, userId);
+
+    if (channel.ownerId !== userId) {
+      throw new ForbiddenException('You are not the owner of this channel');
+    }
+
+    const updatedChannel = await this.prisma.channel.update({
+      data: {
+        users: {
+          disconnect: { id: targetUserId },
+        },
+      },
+      where: { id: channelId },
+    });
+
+    return updatedChannel;
+  }
+
+  async leaveChannel(id: string, userId: string) {
+    const channel = await this.prisma.channel.findFirst({
+      include: {
+        users: { select: this.userSelect },
+      },
+      where: {
+        id,
+        users: { some: { id: userId } },
+      },
+    });
+
+    if (!channel) {
+      throw new NotFoundException('Channel not found or you are not a member of this channel');
+    }
+
+    if (channel.ownerId === userId) {
+      throw new ForbiddenException('You cannot leave a channel you own');
+    }
+    const updatedChannel = await this.prisma.channel.update({
+      data: {
+        users: {
+          disconnect: { id: userId },
+        },
+      },
+      where: { id },
+    });
+    return updatedChannel;
+  }
+
   async remove(id: string, userId: string) {
     const channel = await this.findOne(id, userId);
 
     if (channel.ownerId !== userId) {
       throw new ForbiddenException('You are not the owner of this channel');
+    }
+
+    if (channel.users.length > 1) {
+      throw new ForbiddenException('You cannot delete a channel with other members still in it');
     }
 
     const deletedChannel = await this.prisma.channel.delete({
